requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py
pymongo
requests
pytz
Pillow